requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
python-dotenv>=1.0

# HTMX support
django-htmx>=1.18

# API & HTTP
requests>=2.31

# Async & background tasks (optionnel au début)
celery>=5.3
redis>=5.0

# Export
openpyxl>=3.1
weasyprint>=62.0

# Utils
django-filter>=23.5
django-simple-history>=3.5

# Security & audit
django-auditlog>=2.3
